refactor: clean up styles
