+++ src/app/profile/[username]/page.tsx
@@ -1,8 +1,7 @@
 import { Title } from '@mantine/core'
 import { notFound } from 'next/navigation'
 
-import { ProfileContent } from '@/components/profile/ProfileContent'
-import { CreateWatchlistButton } from '@/components/watchlists/CreateWatchlistButton'
+import { ProfileHeader } from '@/app/profile/components/ProfileHeader'
 import { fetchWithType } from '@/lib/api'
 import { proxyRequestHeaders } from '@/lib/headers'
 import { withBaseURL } from '@/lib/url'
@@ -56,20 +55,17 @@ export default async function ExternalProfilePage({ params }: { params: { userna
   return (
     <div className="flex justify-center">
       <div className="flex min-h-full w-full flex-col space-y-6 p-8 lg:max-w-5xl">
-        <ProfileContent user={userResponse.data}>
+        <ProfileHeader user={userResponse.data} />
+
+        <section>
           <div className="flex items-center justify-between pb-5 pt-10">
-            <Title order={2} className="text-4xl">
+            <Title order={2} className="text-3xl sm:text-4xl">
               Watchlists
             </Title>
-            <CreateWatchlistButton />
           </div>
 
           <div className="flex h-96 flex-col justify-center gap-5 bg-black md:w-full" />
-          <Title order={2} className="pb-5 pt-10 text-4xl">
-            Anime
-          </Title>
-          <div className="flex h-96 flex-col justify-center gap-5 bg-black md:w-full" />
-        </ProfileContent>
+        </section>
       </div>
     </div>
   )
