+++ src/app/profile/page.tsx
@@ -2,15 +2,14 @@ import { Title } from '@mantine/core'
 import { headers } from 'next/headers'
 import { redirect } from 'next/navigation'
 
-import { ProfileContent } from '@/components/profile/ProfileContent'
-import { ArticlesCardsGrid } from '@/components/watchlists/ArticleCardsGrid'
 import { CreateWatchlistButton } from '@/components/watchlists/CreateWatchlistButton'
-import { WatchlistCard } from '@/components/watchlists/WatchlistCard'
 import { getUserFromAuth } from '@/db/users'
 import { fetchWithType } from '@/lib/api'
 import { proxyRequestHeaders } from '@/lib/headers'
 import { createServerClient } from '@/lib/supabase/server'
 import { withBaseURL } from '@/lib/url'
+
+import { ProfileHeader } from './components/ProfileHeader'
 
 import type { WatchlistOverview } from '@/types/watchlists'
 
@@ -43,26 +42,31 @@ export default async function SelfProfilePage() {
   )
 
   return (
-    <div className="flex justify-center">
-      <div className="flex min-h-full w-full flex-col space-y-6 p-8 lg:max-w-5xl">
-        <ProfileContent user={user}>
-          <div className="flex items-center justify-between pb-5 pt-10">
-            <Title order={2} className="text-4xl">
-              Watchlists
-            </Title>
-            <CreateWatchlistButton />
-          </div>
+    <div className="flex min-h-full w-full flex-col justify-center space-y-6 px-6 py-8 sm:px-8 lg:max-w-5xl">
+      <ProfileHeader user={user} />
 
-          <div className="flex h-96 flex-col justify-center gap-5 bg-black md:w-full" />
-          <Title order={2} className="pb-5 pt-10 text-4xl">
+      <section>
+        <div className="flex items-center justify-between pb-5 pt-10">
+          <Title order={2} className="text-3xl sm:text-4xl">
+            Watchlists
+          </Title>
+          <CreateWatchlistButton />
+        </div>
+
+        <div className="flex h-96 flex-col justify-center gap-5 bg-black md:w-full" />
+      </section>
+
+      <section>
+        <div className="flex items-center justify-start pb-5 pt-10">
+          <Title order={2} className="text-3xl sm:text-4xl">
             Anime
           </Title>
-          <div className="flex h-96 flex-col justify-center gap-5 bg-black md:w-full" />
-        </ProfileContent>
-        {/* <pre className="text-wrap text-left">{JSON.stringify(watchlists, null, 2)}</pre> */}
-        <WatchlistCard />
-        <ArticlesCardsGrid />
-      </div>
+        </div>
+        <div className="flex h-96 flex-col justify-center gap-5 bg-black md:w-full" />
+      </section>
+      {/* <pre className="text-wrap text-left">{JSON.stringify(watchlists, null, 2)}</pre> */}
+      {/* <WatchlistCard />
+        <ArticlesCardsGrid /> */}
     </div>
   )
 }
