--- src/app/profile/profile.module.css
@@ -1,6 +1,0 @@
-.test {
-  background-color: --light-dark(
-    var(--mantine-color-gray-0),
-    var(--mantine-color-dark-6)
-  );
-}
