+++ src/app/watchlists/page.tsx
@@ -1,9 +1,9 @@
-import { Group, Title } from '@mantine/core'
+import { Affix, Group, Title } from '@mantine/core'
 
 import { SearchBar } from '@/components/common/Header/SearchBar'
+import { CreateWatchlistButton } from '@/components/watchlists/CreateWatchlistButton'
 import { WatchlistsList } from '@/components/watchlists/WatchlistsList'
 
-import { CreateWatchlistButtonFixed } from './CreateWatchlistButtonFixed'
 import { SortWatchlistsDropdown } from './SortWatchlistsDropdown'
 
 export default function WatchlistsPage() {
@@ -20,15 +20,16 @@ export default function WatchlistsPage() {
           className="w-4/5 md:w-3/5"
         />
         <Group className="my-5 w-5/6 justify-center sm:justify-end md:w-3/4">
-          <CreateWatchlistButtonFixed />
-          <Group>
-            Sort by:
-            <SortWatchlistsDropdown />
-          </Group>
+          Sort by:
+          <SortWatchlistsDropdown />
         </Group>
 
         <WatchlistsList />
       </div>
+
+      <Affix position={{ bottom: 40, right: 40 }} className="flex justify-center">
+        <CreateWatchlistButton size="lg" />
+      </Affix>
     </div>
   )
 }
